Revert "增加语音验证码以及语音通知php,java,python SDKDemo"
