+++ demo/python/SmsSDKDemo.py
@@ -2,11 +2,10 @@
 # encoding:utf-8
 
 import json
-import Qcloud.Sms.sms as SmsSender
-from Qcloud.Sms.voice import VoiceSender
-from Qcloud.Sms.voice import VoicePromptSender
+import Qcloud.Sms.SmsSender as SmsSender
 
-def sms():
+
+def main():
     # 请根据实际 appid 和 appkey 进行开发，以下只作为演示 sdk 使用
     appid = 123456
     appkey = "1234567890abcdef1234567890abcdef"    
@@ -43,30 +42,5 @@ def sms():
     rsp = json.loads(result)
     print result
 
-def voice():
-     # 语音验证码请求
-     voice = VoiceSender(appid=1400024345,appkey="265885e366e02c26b42db642cc522160")
-     result = voice.send(nation_code="86",phone_number="18576696192",playtimes=2,msg="1234",ext="")
-     rsp = json.loads(result)
-     if(int(rsp['result']) != 0):
-         errmsg=rsp['errmsg']
-         print "request failed\n"+errmsg
-     else:
-         print "request success\n"+result
-
-
-    # 语音通知请求
-     voice_promt = VoicePromptSender(appid=1400024345, appkey="265885e366e02c26b42db642cc522160")
-     #note: msg内容，首先需要申请内容模板，通过后才可以发送
-     result = voice_promt.send(nation_code="86", phone_number="18576696192", playtimes=2, msg="你好语音模板", ext="")
-     rsp = json.loads(result)
-     if (rsp['result'] != 0):
-        errmsg = rsp['errmsg']
-        print "request failed\n" + "error code: "+str(rsp['result'])+"\t"+errmsg
-     else:
-        print "request success\n" + result
-
-
 if __name__ == "__main__":
-    sms()
-    voice()
+    main()
